Fix for gcc 6.2.1 error: left shift of negative value [-Werror=shift-negative-value]

+++ groups/bdl/bdlb/bdlb_bitmaskutil.h
@@ -252,7 +252,7 @@ bsl::uint32_t BitMaskUtil::ge(int index)
 
     return BSLS_PERFORMANCEHINT_PREDICT_LIKELY(
                                    index < static_cast<int>(k_BITS_PER_UINT32))
-           ? (~0 << index)
+           ? (~static_cast<uint32_t>(0) << index)
            : 0;
 }
